requirements/development.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.4
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.93.3
orjson==3.10.18
packaging==24.2
pgvector==0.4.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
zstandard==0.23.0
